refactor(orchestrator): hoist get_git_head + TRANSIENT_EXCEPTIONS to shared module
Both symbols were duplicated byte-for-byte in service.py and runner.py
(service imports runner to avoid a cycle). Move them into a new
dependency-free sibling module _common.py that both import, eliminating
the drift risk. service.py and runner.py re-export the names into their
namespaces, preserving every existing import/patch target
(service.get_git_head, service.TRANSIENT_EXCEPTIONS).
